requirements.txt
boto3
Celery
chardet
django==5.2.7
django-cors-headers
django-silk
django-storages
django-tenants
djangorestframework==3.16.1
djangorestframework-simplejwt
djoser
drf-spectacular
flower
google
google-generativeai
imagehash
isort==7.0.0
matplotlib
numpy
opencv-contrib-python
opencv-python
openpyxl
pandas
pdf2image
pillow
poppler-utils
PyPDF2
python-dateutil
python-dotenv
pymupdf
pytesseract
redis[hiredis]
regex
reportlab
requests
scikit-image
scikit-learn
scipy
torch
torchvision
tqdm
tenacity
playwright
pypdf
easyocr
psycopg2
whitenoise